Updated Works component: the title has become a component.
