+++ src/scenes/Portfolio/components/Works/components/List/List.module.scss
@@ -1,0 +1,72 @@
+@import '@styles/variables/all';
+@import '@styles/mixins/all';
+
+.container {
+}
+
+.buttonWrapper {
+  display: flex;
+  justify-content: center;
+  margin-top: 5.71rem;
+
+  @include media-breakpoint-down(xs) {
+    margin-top: 3.34rem;
+  }
+
+  a {
+    @media screen and (max-width: 1200px) {
+      font-size: 14px;
+    }
+  }
+}
+
+.button {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0 1rem;
+  font-size: 1rem;
+  font-weight: bold;
+  line-height: 2.43em;
+  letter-spacing: 0.71em;
+  text-decoration: none;
+  color: $yellow-color;
+
+  &:before {
+    content: '';
+    display: block;
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    background: $main-color;
+    z-index: -1;
+    transform: scaleY(0);
+    transition: transform 0.2s;
+    transform-origin: bottom;
+  }
+
+  svg {
+    display: block;
+    width: 1.14em;
+    height: auto;
+    margin-left: 0.36em;
+    fill: currentColor;
+  }
+
+  &:hover {
+    color: #ffffff;
+
+    svg {
+      fill: currentColor;
+    }
+
+    @include media-breakpoint-up(md) {
+      &:before {
+        transform: scaleY(1);
+      }
+    }
+  }
+}
