+++ src/scenes/Portfolio/components/Works/components/List/index.js
@@ -1,0 +1,3 @@
+import List from './List';
+
+export default List;
