+++ src/scenes/Portfolio/components/Works/components/Title/Title.module.scss
@@ -1,0 +1,78 @@
+@import '@styles/variables/all';
+@import '@styles/mixins/all';
+
+.container {
+  font-family: $font-ratiomodern;
+  color: $white-color;
+  font-weight: 400;
+  text-align: center;
+  display: inline-flex;
+  position: relative;
+  font-size: 4.58rem;
+  line-height: 1.598em;
+  margin-bottom: 5.57rem;
+
+  @include media-breakpoint-down(md) {
+    font-size: 3.67rem;
+    margin-bottom: 5rem;
+    line-height: 1.46em;
+  }
+
+  @include media-breakpoint-down(sm) {
+    font-size: 4.5rem;
+    line-height: 1.1em;
+    margin-bottom: 2.5rem;
+  }
+
+  @include media-breakpoint-down(xs) {
+    font-size: 2.83rem;
+    margin-bottom: 2.5rem;
+  }
+}
+
+.link {
+  display: block;
+  position: absolute;
+  left: 100%;
+  bottom: 100%;
+  transform: translate(6%, 25%);
+  line-height: 0;
+  width: 2em;
+
+  @include media-breakpoint-down(md) {
+    display: none;
+  }
+
+  img {
+    width: 100%;
+  }
+
+  &:hover {
+    .image,
+    .icon {
+      animation-play-state: running;
+    }
+  }
+}
+
+.image {
+  animation: dribbleAnimationRotate 10s infinite linear;
+  animation-play-state: paused;
+}
+
+.icon {
+  position: absolute;
+  max-width: 3.18rem;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+@keyframes dribbleAnimationRotate {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
