+++ src/scenes/Portfolio/components/Works/components/Title/index.js
@@ -1,0 +1,3 @@
+import Title from './Title';
+
+export default Title;
